config(ticdc): split config
